Optional force run on single CPU core for Windows.
New CVAR to select on which single CPU core the engine should run. This
is used to speed up the engine in some cases when using NVIDIA cards.
The new CVAR is:
\com_singlecore
Values: 0 = Normal (do not force single core)
1 = use CPU core no. 1
2 = use CPU core no. 2
Etc...
If user CPU core is not found, and value is > 1 the last CPU core will
be forced.