Fixed dmaHD entering an infinite loop on certain maps (fixes FrozenSand/UrbanTerror42#34)
In certain cases (sound/misc/windfly.wav in ut4_kingsgambit_a9.pk3, for example), dmaHD_NormalizeSamplePosition will get a "samples" value of 0, and it will enter an infinite loop. This is happening because the "data" chunk of a wav file has a size of 0. According to ioquake/ioq3@ae39051f, these are legal, so I didn't change the behaviour of the wav file loader.