Add sv_extraPure and sv_extraPaks cvars.
When set to 1, the server will exclude unreferenced paks from the pure
paks list. Unpure servers are unaffected.

There are two main uses for this option:

 - Servers having so many maps that they overflow the server info string
   can be pure again.

 - Conflicts between maps can be avoided, since only the map and
   core game paks will be loaded by the client.

Clients connecting to an extra pure server will only have the core game
paks and the server-referenced paks (ie the map) in their search path,
so no conflict can exist with other maps.

On the other hand, this will prevent content from other maps to be
loaded (such as funstuff). The server will act as if all other maps were
not present.

To counter this, another sv_extraPaks cvar is added. It's a
space-separated list of paks that will not be excluded from the pure pak
list, eg, sv_extraPaks "map1 map2". The order doesn't matter.