Add fs_reorderPaks cvar.
When set to 1 (the default) and a map is loaded, this will reorder paks
on according to the following order:

 - Pak containing the map being loaded (unless it's a zUrT pak)
 - All zUrT paks in their original order
 - All remaining paks and dirs in their original order

The intended effect is to reduce conflicts between loaded maps:

 - The current map always comes first, so it can not be overriden by
   anything else.

 - The core game paks come next, so they don't get overriden by other
   maps that change default textures, sounds, weapons, etc.

 - Finally, the other paks are included, which allows use of eg,
   funstuff.

When applied to a pure server, this will affect connected clients since
pure clients reorder their paks using the server's order. So clients
without this cvar will benefit from the change.

When applied to a client, reordering will also be done, including
when connecting to a pure server. This means that the client will likely
have a different search order than the server which deviates from the
standard pure behavior of following to the server order. This doesn't
break the pureness of the client though.