Copy server address before using it
We can't assume that no calls to `Cmd_TokenizeString` will be made in
the middle of `CL_Connect_f` which would overwrite the command buffer
where the server address is still stored.

In fact `Cmd_TokenizeString` is actually called by
`FS_PureServerSetReferencedPaks` which is itself called from
`CL_Disconnect` but this doesn't cause any problems since the parsed
string is always empty.