Tweak up some sound limits
There is a limit on the number of concurrent playing channels for each
single sound.

The limit has always been set to 8 for your own sounds (and 4 for other
players') but there was a bug in the original Quake3 source that made
this limit unenforced.

Since ioUrbanTerror never fixed this bug, it was never found that firing
some weapons like the LR300 would hit that limit.

Raising both limits to 16 fixes this issue for good.

Additionaly, there is also a minimum delay after which a sound that
started playing can be played again.

This limit was set to 50ms which seems to be regularly reached during
normal game play, so let's lower than to 20 ms instead.

I've also added warnings so that we always know why and when a sound
isn't played.

Fixes #19